Introduce oracle module
