+++ cfd_protocol/src/oracle/bip340_hash.rs
@@ -1,0 +1,15 @@
+use bdk::bitcoin::hashes::*;
+
+sha256t_hash_newtype!(
+    Bip340Hash,
+    Bip340HashTag,
+    BIP340_MIDSTATE,
+    64,
+    doc = "bip340 hash",
+    true
+);
+
+const BIP340_MIDSTATE: [u8; 32] = [
+    0x9c, 0xec, 0xba, 0x11, 0x23, 0x92, 0x53, 0x81, 0x11, 0x67, 0x91, 0x12, 0xd1, 0x62, 0x7e, 0x0f,
+    0x97, 0xc8, 0x75, 0x50, 0x00, 0x3c, 0xc7, 0x65, 0x90, 0xf6, 0x11, 0x64, 0x33, 0xe9, 0xb6, 0x6a,
+];
